+++ src/io/unix/sync.rs
@@ -52,7 +52,8 @@ pub(crate) struct File {
 #[derive(Debug)]
 struct FileInner {
     std_file: StdFile,
-    size: AtomicU64
+    size: AtomicU64,
+    synced_size: AtomicU64
 }
 
 #[derive(PartialEq, Eq)]
@@ -65,6 +66,12 @@ enum LockAcquisitionResult {
 impl File {
     pub(crate) fn size(&self) -> u64 {
         self.inner.size.load(Ordering::SeqCst)
+    }
+    pub(crate) fn synced_size(&self) -> u64 {
+        self.inner.synced_size.load(Ordering::SeqCst)
+    }
+    pub(crate) fn dirty_bytes(&self) -> u64 {
+        self.size() - self.synced_size()
     }
 
     pub(crate) async fn write_append_writable_data<R: Send + 'static>(
@@ -156,7 +163,14 @@ impl File {
 
     pub(crate) async fn fsyncdata(&self) -> IOResult<()> {
         let file_inner = self.inner.clone();
-        Self::background_sync_call(move || file_inner.std_file.sync_all()).await
+        let size = self.size();
+        Self::background_sync_call(
+            move || {
+               file_inner.std_file.sync_all()?;
+               file_inner.synced_size.fetch_max(size, Ordering::SeqCst);
+               Ok(())
+            }
+        ).await
     }
 
     pub(crate) fn created_at(&self) -> IOResult<SystemTime> {
@@ -172,6 +186,11 @@ impl File {
     #[cfg(feature = "async-io-rio")]
     pub(super) fn file_size_append(&self, len: u64) -> u64 {
         self.inner.size.fetch_add(len, Ordering::SeqCst)
+    }
+
+    #[cfg(feature = "async-io-rio")]
+    pub(super) fn set_synced_size(&self, synced_size: u64) {
+        self.inner.synced_size.fetch_max(synced_size, Ordering::SeqCst);
     }
 
     fn advisory_write_lock_file(fd: i32) -> LockAcquisitionResult {
@@ -235,13 +254,15 @@ impl File {
 
     async fn from_tokio_file(file: TokioFile) -> IOResult<Self> {
         let size = file.metadata().await?.len();
+        let synced_size = AtomicU64::new(size);
         let size = AtomicU64::new(size);
         let std_file = file.try_into_std().expect("tokio file into std");
 
         let file = Self {
             inner: Arc::new(FileInner { 
                 std_file, 
-                size 
+                size,
+                synced_size
             })
         };
         Ok(file)
@@ -258,6 +279,9 @@ impl super::super::FileTrait for File {
     }
     fn created_at(&self) -> IOResult<SystemTime> {
         self.created_at()
+    }
+    fn dirty_bytes(&self) -> u64 {
+        self.dirty_bytes()
     }
 
     async fn write_append_writable_data<R: Send + 'static>(&self, c: impl WritableDataCreator<R>) -> IOResult<R> {