+++ CHANGELOG.md
@@ -5,9 +5,12 @@ Pearl changelog
 ## [Unreleased]
 #### Added
 - Checksum validation in `Entry::load_data` (#274)
+- Add fsync to header writing in blob (#243)
+- Add periodic fsync (#234)
 
 #### Changed
-
+- Reduced internal strucutre size in memory: bloom filter `Config` and `FileName` (#260)
+- Serialize/Deserialize implementation removed from `Record` struct (#282)
 
 #### Fixed
 - Use `dep:<crate_name>` syntax in feature declaration to avoid unnecessary feature flags (#272)